+++ classes/Controllers/PostTypes.php
@@ -28,6 +28,11 @@ class PostTypes extends Controller {
 		add_action( 'init', [ $this, 'register_post_types' ] );
 		add_action( 'save_post_popup', [ $this, 'save_post' ], 10, 3 );
 		add_filter( 'post_updated_messages', [ $this, 'updated_messages' ] );
+
+		// Control block editor usage on per popup post type basis.
+		add_filter( 'use_block_editor_for_post_type', [ $this, 'use_block_editor_for_post_type' ], 10, 2 );
+		// Blanket disable block editor for popup post types if setting is enabled.
+		add_action( 'replace_editor', [ $this, 'replace_editor' ], 10, 2 );
 	}
 
 	/**
@@ -302,7 +307,7 @@ class PostTypes extends Controller {
 	 * @return void
 	 */
 	public function register_popup_category_tax() {
-		if ( $this->container->get_option( 'disable_popup_category_tag', false ) ) {
+		if ( pum_get_option( 'disable_popup_category_tag', false ) ) {
 			return;
 		}
 
@@ -345,7 +350,7 @@ class PostTypes extends Controller {
 	 * @return void
 	 */
 	public function register_popup_tag_tax() {
-		if ( $this->container->get_option( 'disable_popup_category_tag', false ) ) {
+		if ( pum_get_option( 'disable_popup_category_tag', false ) ) {
 			return;
 		}
 
@@ -509,5 +514,48 @@ class PostTypes extends Controller {
 		}
 
 		return $messages;
+	}
+
+	/**
+	 * Control whether to use block editor for popup post types.
+	 *
+	 * @param bool   $use_block_editor Whether to use the block editor.
+	 * @param string $post_type        The post type.
+	 * @return bool Whether to use the block editor for this post type.
+	 */
+	public function use_block_editor_for_post_type( $use_block_editor, $post_type ) {
+		// Only control our post types.
+		if ( ! in_array( $post_type, [ 'popup', 'popup_theme' ], true ) ) {
+			return $use_block_editor;
+		}
+
+		// If classic editor is enabled, disable block editor.
+		if ( pum_get_option( 'enable_classic_editor', false ) ) {
+			return false;
+		}
+
+		// Otherwise use block editor.
+		return true;
+	}
+
+	/**
+	 * Replace the editor interface when classic editor is forced.
+	 *
+	 * @param bool    $replace   Whether to replace the editor.
+	 * @param WP_Post $post      The post object.
+	 * @return void
+	 */
+	public function replace_editor( $replace, $post ) {
+		// Only handle our post types.
+		if ( ! in_array( $post->post_type, [ 'popup', 'popup_theme' ], true ) ) {
+			return;
+		}
+
+		// If classic editor is enabled, ensure we're using classic interface.
+		if ( pum_get_option( 'enable_classic_editor', false ) ) {
+			// Remove block editor assets if they were enqueued.
+			remove_action( 'admin_enqueue_scripts', 'wp_enqueue_editor' );
+			remove_action( 'admin_footer', 'wp_enqueue_editor' );
+		}
 	}
 }